+++ CHANGELOG.md
@@ -2,9 +2,20 @@
 
 ## Unreleased
 
+### Fixes
+
+- Don't start the SDK inside Xcode preview (#4601)
+
+### Improvements
+
+- Add native SDK information in the replay option event (#4663)
+- Add error logging for invalid `cacheDirectoryPath` (#4693)
+
 ### Features
 
+- SwiftUI time for initial display and time for full display (#4596)
 - Add protocol for custom screenName for UIViewControllers (#4646)
+- Allow hybrid SDK to set replay options tags information (#4710)
 - Add threshold to always log fatal logs (#4707)
 
 ### Internal
@@ -19,7 +30,6 @@
 - Memory growth issue in profiler (#4682)
 - Replace occurences of `strncpy` with `strlcpy` (#4636)
 - Fix span recording for `NSFileManager.createFileAtPath` starting with iOS 18, macOS 15 and tvOS 18. This feature is experimental and must be enabled by setting the option `experimental.enableFileManagerSwizzling` to `true` (#4634)
-
 
 ### Internal